+++ server/src/main/java/org/allaymc/server/block/component/BlockQuartzBlockBaseComponentImpl.java
@@ -1,0 +1,23 @@
+package org.allaymc.server.block.component;
+
+import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.dto.PlayerInteractInfo;
+import org.allaymc.api.block.type.BlockState;
+import org.allaymc.api.block.type.BlockType;
+import org.allaymc.api.world.Dimension;
+import org.joml.Vector3ic;
+
+/**
+ * @author IWareQ
+ */
+public class BlockQuartzBlockBaseComponentImpl extends BlockBaseComponentImpl {
+    public BlockQuartzBlockBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
+        super(blockType);
+    }
+
+    @Override
+    public boolean place(Dimension dimension, BlockState blockState, Vector3ic placeBlockPos, PlayerInteractInfo placementInfo) {
+        dimension.setBlockState(placeBlockPos.x(), placeBlockPos.y(), placeBlockPos.z(), blockState, placementInfo);
+        return true;
+    }
+}